+++ server/photomory/src/main/java/com/example/photomory/service/OurAlbumService.java
@@ -25,7 +25,7 @@ public class OurAlbumService {
     }
 
     @Transactional(readOnly = true)
-    public OurAlbumResponseDto getAlbumDetails(Integer albumId, Integer requestUserId) {
+    public OurAlbumResponseDto getAlbumDetails(Integer albumId, Long requestUserId) {
         // 1. 앨범 조회 (없으면 404)
         Album album = albumRepository.findById(albumId)
                 .orElseThrow(() -> new EntityNotFoundException("앨범을 찾을 수 없습니다."));
@@ -34,11 +34,11 @@ public class OurAlbumService {
         if (album.getPosts() == null || album.getPosts().isEmpty()) {
             throw new EntityNotFoundException("앨범에 게시물이 없습니다.");
         }
-        Integer albumOwnerUserId = album.getPosts().get(0).getUser().getUserId();
+        Long albumOwnerUserId = album.getPosts().get(0).getUser().getUserId();
 
         // 3. 요청 유저와 앨범 소유자가 친구인지 확인 (없으면 403)
         boolean areFriends = friendRepository.existsByFromUserIdAndToUserId(
-                requestUserId.longValue(), albumOwnerUserId.longValue()
+                requestUserId, albumOwnerUserId
         );
         if (!areFriends) {
             throw new SecurityException("친구가 아니면 접근할 수 없습니다.");
@@ -46,7 +46,7 @@ public class OurAlbumService {
 
         // 4. 요청 유저가 앨범의 그룹 멤버인지 확인 (없으면 403)
         boolean isGroupMember = albumMembersRepository.existsByUserEntityUserIdAndMyAlbumMyalbumId(
-                requestUserId, album.getMyAlbum().getMyalbumId()
+                requestUserId.intValue(), album.getMyAlbum().getMyalbumId()
         );
         if (!isGroupMember) {
             throw new SecurityException("그룹 멤버만 접근할 수 있습니다.");
